+++ src/components/ConfirmationValues.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import '../styles/confirmationValues.scss';
+import '../styles/Confirmation/confirmationValues.scss';
 
 class ConfirmationValues extends Component {
     render() {
@@ -7,7 +7,7 @@ class ConfirmationValues extends Component {
         return (
             <section className="confirmation-infos">
                 <div className="confirmation-values">
-                    <h4>{cardValues.cardNumber}</h4>
+                    <h4>{cardValues.cardNumber.replace(/.(?=.{4})/g, '*')}</h4>
                     <h4>{cardValues.personName}</h4>
                     <h4>{cardValues.expire}</h4>
                 </div>
