+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "react-dom";
 import Root from './root';
-import configureStore from './configureStore';
+import configureStore from './services/configureStore';
 import './styles/index.scss';
 
 // colocamos toda a lógica de criar um store em outro arquivo
