+++ src/services/configureStore.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from "redux";
 import thunk from 'redux-thunk';
-import reducers from "./reducers";
+import reducers from "../reducers";
 import { loadState, saveState } from './localStorage';
 
 const configureStore = () => {
