+++ src/styles/Payment/payment.scss
@@ -1,8 +1,12 @@
-@import 'button';
+@import '../button';
 
 .form {
     height: auto;
     width: calc(100% - 20px);
+
+    @media (min-width: 1024px) {
+        width: 40%;
+    }
 
     > h6 {
         color: #999;
