+++ src/styles/header.scss
@@ -1,0 +1,26 @@
+header {
+    background: #FFF;
+    box-sizing: border-box;
+    box-shadow: 1px 1px 5px 0 rgba(0,0,29,0.22);
+    height: 50px;
+    width: 100%;
+
+    @media (min-width: 1024px) {
+        width: 40%;
+        margin: 0 auto;
+    }
+}
+
+nav {
+    height: 100%;
+    width: 100%;
+}
+
+ul {
+    align-items: center;
+    display: flex;
+    flex-flow: row nowrap;
+    height: 100%;
+    justify-content: space-around;
+    width: 100%;
+}
